requirements.txt
# 基础依赖
PySide6>=6.4.0
opencv-python>=4.5.0
numpy>=1.20.0
Pillow>=9.0.0
